requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-browser-reload==1.12.1
django-environ==0.11.2
factory-boy==3.3.0
Faker==25.2.0
pillow==10.3.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
